+++ solr/modules/language-models/src/java/org/apache/solr/languagemodels/LanguageModelException.java
@@ -14,6 +14,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package org.apache.solr.languagemodels;
 
-/** Contains update request processor related classes. */
-package org.apache.solr.languagemodels.textvectorisation.update.processor;
+public class LanguageModelException extends RuntimeException {
+
+  private static final long serialVersionUID = 1L;
+
+  public LanguageModelException(String message) {
+    super(message);
+  }
+
+  public LanguageModelException(String message, Exception cause) {
+    super(message, cause);
+  }
+}
