+++ solr/modules/language-models/src/java/org/apache/solr/languagemodels/update/processor/package-info.java
@@ -15,5 +15,5 @@
  * limitations under the License.
  */
 
-/** Contains the {@link org.apache.solr.rest.ManagedResource} that encapsulate the model stores. */
-package org.apache.solr.languagemodels.textvectorisation.store.rest;
+/** Contains update request processor related classes. */
+package org.apache.solr.languagemodels.update.processor;
