+++ solr/modules/language-models/src/test/org/apache/solr/languagemodels/textvectorisation/store/rest/TestManagedLanguageModelStoreInitialization.java
@@ -17,6 +17,7 @@
 package org.apache.solr.languagemodels.textvectorisation.store.rest;
 
 import org.apache.solr.languagemodels.TestLanguageModelBase;
+import org.apache.solr.languagemodels.store.rest.ManagedTextToVectorModelStore;
 import org.junit.After;
 import org.junit.Test;
